requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
requests==2.31.0
Pillow==9.5.0
supabase==2.4.5
langgraph==0.0.15
langchain-core==0.1.25
typing-extensions==4.5.0
PyPDF2==3.0.1
pdfplumber==0.10.4
pytesseract==0.3.10
opencv-python-headless==4.8.1.78